backend/requirements.txt
fastapi
uvicorn
pyserial
pandas
tensorflow
numpy
pyttsx3
python-dotenv
gunicorn
